+++ fabric/src/main/java/de/maxhenkel/voicechat/FabricVoicechatMod.java
@@ -2,6 +2,7 @@ package de.maxhenkel.voicechat;
 
 import de.maxhenkel.configbuilder.ConfigBuilder;
 import de.maxhenkel.voicechat.config.FabricServerConfig;
+import de.maxhenkel.voicechat.integration.ViaVersionCompatibility;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.loader.api.FabricLoader;
 
@@ -12,6 +13,8 @@ public class FabricVoicechatMod extends Voicechat implements ModInitializer {
         SERVER_CONFIG = ConfigBuilder.build(FabricLoader.getInstance().getConfigDir().resolve(MODID).resolve("voicechat-server.properties"), true, FabricServerConfig::new);
 
         initialize();
+
+        ViaVersionCompatibility.register();
     }
 
 }